requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
audioread==3.0.1
azure-cognitiveservices-speech==1.44.0
azure-core==1.34.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-storage-blob==12.27.1
blinker==1.9.0
boto3==1.41.0
botocore==1.41.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docx==0.2.4
fastapi==0.121.2
fastuuid==0.14.0
ffmpeg==1.4
filelock==3.18.0
fonttools==4.58.0
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain-core==0.3.63
langsmith==0.3.45
lazy_loader==0.4
librosa==0.11.0
litellm==1.80.0
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.7.0
narwhals==1.38.2
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
opencv-python==4.11.0.86
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
propcache==0.4.1
protobuf==6.30.2
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pyppeteer==2.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
s3transfer==0.14.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.2
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.22.1
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==10.4
whisper==1.1.10
yarl==1.22.0
zipp==3.22.0
zstandard==0.23.0